+++ src/frontend/apps/impress/src/features/docs/doc-management/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './useCollaboration';
 export * from './useTrans';
+export * from './useCopyDocLink';