+++ src/frontend/apps/impress/src/hook/index.ts
@@ -1,1 +1,2 @@
 export * from './useDate';
+export * from './useClipboard';